rename dither to pattern_expansion
