+++ observation_portal/requestgroups/pattern_expansion.py
